+++ src/shared/tdoc-commons/model/repository/sql-mytbdb-destination.config.ts
@@ -202,6 +202,14 @@ export class SqlMytbDbDestinationConfig {
                 filterFields: ['tif.if_id'],
                 action: AdapterFilterActions.IN_NUMBER
             },
+            'initial_s': {
+                selectSql: 'SELECT COUNT(*) as count, ' +
+                    ' SUBSTR(UPPER(d_name), 1, 1) as value ' +
+                    'FROM destination ' +
+                    'WHERE LENGTH(d_name) > 0 ' +
+                    'GROUP BY SUBSTR(UPPER(d_name), 1, 1)' +
+                    'ORDER BY value',
+            },
             'keywords_txt': {
                 selectSql: 'SELECT 0 AS count, ' +
                     '  kw_name AS value ' +
@@ -541,6 +549,7 @@ export class SqlMytbDbDestinationConfig {
             loc_id_is: 'destination.l_id',
             loc_lochirarchie_ids_txt: 'location.l_id',
             l_lochirarchietxt: 'location.l_name',
+            initial_s: 'SUBSTR(UPPER(d_name), 1, 1)',
             html: 'CONCAT(d_name, " ", COALESCE(d_desc_gebiet, ""), " ", l_name)'
         },
         writeMapping: {