+++ src/shared/tdoc-commons/model/repository/sql-mytbdb-news.config.ts
@@ -245,6 +245,9 @@ export class SqlMytbDbNewsConfig {
                 selectField: 'CONCAT("DONE", (n_date IS NOT NULL))',
                 orderBy: 'value asc'
             },
+            'initial_s': {
+                noFacet: true
+            },
             'keywords_txt': {
                 noFacet: true
             },
@@ -386,6 +389,7 @@ export class SqlMytbDbNewsConfig {
             route_id_is: '"666dummy999"',
             loc_lochirarchie_ids_txt: '"666dummy999"',
             l_lochirarchietxt: 'location.l_name',
+            initial_s: '"666dummy999"',
             html: 'CONCAT(n_headline, " ", COALESCE(n_message,""))'
         },
         writeMapping: {